fixed update issue with system parameters

+++ HOPSAN++/HopsanGUI/Widgets/SystemParametersWidget.h
@@ -59,7 +59,7 @@ public slots:
     void typeHasChanged(QString newType);
 
 protected:
-    size_t mRow, mColumn;
+    int mRow, mColumn;
     SystemParameterTableWidget *mParent;
 };
